pyproject.toml
[project]
name = "exam-pdf-parser"
version = "1.0.0"
description = "Parse exam PDFs using 3-layer architecture: Document Parsers + LLM Structuring + Validation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "bsh998", email = "bsh998@snu.ac.kr" },
]
keywords = ["pdf", "parser", "exam", "ocr", "llm", "korean", "english"]

dependencies = [
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "easyocr>=1.7.0",
]
ocr-heavy = [
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "surya-ocr>=0.4.0",
]
ocr-transformer = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
ocr-deepseek = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "einops>=0.7.0",
    "addict>=2.4.0",
    "easydict>=1.10",
]
doc-marker = [
    "marker-pdf>=1.0.0",
]
doc-mineru = [
    "mineru>=1.0.0",
]
doc-docling = [
    "docling>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
benchmark = [
    "matplotlib>=3.7.0",
    "tabulate>=0.9.0",
]
all = [
    "exam-pdf-parser[doc-docling]",
    "exam-pdf-parser[ocr]",
    "exam-pdf-parser[ocr-transformer]",
    "exam-pdf-parser[dev]",
    "exam-pdf-parser[benchmark]",
]

[project.scripts]
exam-parser = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"scripts/validate.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# mineru requires Pillow>=11; marker-pdf, surya-ocr require Pillow<11
conflicts = [
    [
        { extra = "doc-marker" },
        { extra = "doc-mineru" },
    ],
    [
        { extra = "ocr-heavy" },
        { extra = "doc-mineru" },
    ],
]
